conn pool: unify uint32_t in connection pool clients and max_request_per_connection (#39275)
1. `max_requests_per_connection` is `uint32Value` in the http options
and cluster configs, so no need to use `uint64_t`
2. clean up for both HTTP and TCP active clients in the connection pool
to avoid `uint64_t` for any limits and the base
`Envoy::ConnectionPool::ActiveClient`are all `uint32_t`.
3. correct comments and other changes for the default.

+++ source/server/admin/admin.h
@@ -220,7 +220,7 @@ public:
   void closeSocket() override;
   void addListenerToHandler(Network::ConnectionHandler* handler) override;
 
-  uint64_t maxRequestsPerConnection() const override { return 0; }
+  uint32_t maxRequestsPerConnection() const override { return 0; }
   const HttpConnectionManagerProto::ProxyStatusConfig* proxyStatusConfig() const override {
     return proxy_status_config_.get();
   }
